+++ app/src/main/java/com/farware/recipesaver/feature_recipe/data/entities/RecipeIngredientEntity.kt
@@ -26,7 +26,7 @@ data class RecipeIngredientEntity(
     val recipeIngredientId: Long? = null,
     val recipeId: Long,
     val ingredientId: Long,
-    val measureId: Int,
+    val measureId: Long,
     val amount: String,
     val measure: String,
     val ingredient: String
