+++ app/src/main/java/com/farware/recipesaver/feature_recipe/domain/model/recipe/RecipeIngredient.kt
@@ -6,7 +6,7 @@ data class RecipeIngredient(
     val recipeIngredientId: Long? = null,
     val recipeId: Long,
     val ingredientId: Long,
-    val measureId: Int,
+    val measureId: Long,
     val amount: String,
     val measure: String,
     val ingredient: String
