+++ app/src/main/java/com/farware/recipesaver/feature_recipe/domain/model/recipe/relations/RecipeIngredientWithIngredient.kt
@@ -4,7 +4,7 @@ data class RecipeIngredientWithIngredient(
     val recipeIngredientId: Long? = null,
     val recipeId: Long,
     val ingredientId: Long,
-    val measureId: Int,
+    val measureId: Long,
     val amount: String,
     val measure: String,
     val ingredient: String,
